pyproject.toml
[tool.poetry]
name = "jembe-auth-demo"
version = "0.1.0"
description = "Simple user auth and user management implmented using jembe framework"
authors = ["Predrag Peranović <predrag.peranovic@blokkod.me>"]
license = "AGPLv3"

[tool.poetry.dependencies]
python = "^3.8"
Flask-SQLAlchemy = "^2.4.4"
Flask-Migrate = "^2.7.0"
python-dotenv = "^0.15.0"
WTForms = {extras = ["email"], version = "^2.3.3"}
WTForms-SQLAlchemy = "^0.2"
Flask-Babel = "^2.0.0"
email-validator = "^1.1.2"
Flask-SeaSurf = "^0.3.0"
Flask-Session = "^0.3.2"
Flask-Login = "^0.5.0"

[tool.poetry.dev-dependencies]
jembe = {path = "../jembe", develop=true}
black = "^20.8b1"
mypy = "^0.812"
pytest = "^6.2.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
